fix: the directory name has been changed and minor errors have been fixed
