+++ CMakeLists.txt
@@ -27,7 +27,7 @@ set(DEPLEX_LIB_DIR ${CMAKE_BINARY_DIR}/lib)
 add_subdirectory(external)
 add_subdirectory(libs)
 add_subdirectory(cpp)
-add_subdirectory(benchmark-artifact/cpp-benchmark)
+add_subdirectory(benchmark/cpp-benchmark)
 if (${BUILD_EXAMPLES})
     add_subdirectory(examples)
 endif()
