+++ benchmark/data/color/000004415600.png
