+++ benchmark/data/depth/000004415622.png
